feat(curriculum): add storage service and improve validation
- Implement CurriculumImportStorage service for caching curriculum JSON
- Change DurationYears from int? to double? for more precise duration
- Add hash-based caching to avoid redundant AI extraction
- Improve JSON parsing with case-insensitive options
- Update validation rules for credits and term numbers
- Refactor API routes for better organization
- Add storage service to test dependencies

+++ src/RogueLearn.User.Application/Models/CurriculumImportSchema.cs
@@ -41,7 +41,7 @@ public class CurriculumProgramData
     
     public int? TotalCredits { get; set; }
     
-    public int? DurationYears { get; set; }
+    public double? DurationYears { get; set; }
 }
 
 /// <summary>
